refactor(buildsource): extract L2 include-seeding into l2_seed module
The L2 include-seeding helpers (derive_l2_include_dirs, seed_l2_includes) grew
inline.py past the 2000-line hard cap (AI-readiness gate). They are a cohesive,
self-contained reader over collect_inline_pack, so move them into a sibling
abicheck/buildsource/l2_seed.py (inline.py drops to ~1820 lines).

l2_seed imports its collection deps one-way from inline; inline re-exports the
two helpers via a lazy module __getattr__ so the historical
 paths and tests keep working without a
static inline -> l2_seed edge that would re-introduce an import cycle (same shim
pattern as cli_buildsource). The CLI callers import directly from l2_seed so the
call sites keep a concrete callable type for mypy. Behaviour is unchanged; the
existing tests cover the moved code.
